Driver: Fix a -Wshadow issue from r237091
The MachO toolchain has an isTargetIOSBased method, but it isn't
virtual so it isn't very meaningful to call it. After thinking about
this, I guess that putting this logic in the MachO class is a bit of a
layering violation anyway. Do this more like how we handle
AddLinkRuntimeLibArgs instead.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@237095 91177308-0d34-0410-b5e6-96231b3b80d8